Disable the PETSc reference counting check for 3.20 and newer.
The 3.20 release completely rewrote the way logging works in PETSc and this old
approach now crashes. I know more-or-less how to fix it (there's a relevant
PETSc example) but I don't have time to do it this week - for now lets just
disable this extra check with the latest PETSc version.