Tensor<k, dim, Number::unroll(): fix copy statement
The task for the rank-1 case is to copy all values from the internal
values array to the iterator range, so let us just use
```
  std::copy(std::begin(values), std::end(values), begin);
```
for this. The problem with the `&values[dim]` statement is the fact that
values[dim] is an out-of-bounds access and the address of if is
undefined.