Fix unintended terminate in gmsh-API version of read_msh
The gmsh-api version of GridIn::read_msh uses a throw statement without
specifing an exception. The intention is cleary to throw any exception
but the standard specifies that doing this outside of a catch block
should lead to std::terminate being called. Set the boundary_id directly,
without using throw, and add a test that covers this part of the function.