Handle zero contour aligning with faces in MeshClassifier.
Change logic in MeshClassifier to handle the situation described in
issue #16888. Add a new type to the enum LocationToLevelSet that
describes that the zero contour is exactly aligned with a face.
Change the logic in MeshClassifier to set the "inside" cell sharing
an aligned face to intersected and the other cell to outside.

Co-authored-by: Magdalena Schreter <magdalena.schreter@tum.de>