Add ReferenceCell::max_n_vertices<dim>().
Like the other max functions, this helps us semantically disambiguate
between a maximum value and the hypercube value.

I had to add more preprocessor checks to work around Windows Server
2019's MSVC's problems with boost. I'm not sure when GitHub will start
removing support for that OS, but it may be soon: it entered extended
support a year ago [1]. I think we should just leave these preprocessor
defines until we retire support for that OS, which may be after the next
release.

[1] https://learn.microsoft.com/en-us/lifecycle/products/windows-server-2019