Break the cycle of header includes between template_constraints.h and complex_overloads.h.
We do not ever directly include complex_overloads.h, from anywhere except
template_constraints.h. As a consequence, the include of
template_constraints.h in complex_overloads.h always expands into nothing
(because of the header guard in template_constraints.h) and it is safe
to remove the include in complex_overloads.h.