MappingP1: only compute the covariant once.
I did some more profiling and saw that, in the same load vector
benchmark, I was computing the covariant about 10 places per cell - its
easy enough to just compute it once and store it with the rest of the
transformation.

I elected to unconditionally compute the covariant (and determinant of
the contravariant in the last PR) since they are constant on each cell
and needed most of the time.