fix nvcc work-around in petsc_matrix_base.h
Compilation fails for me even with the workaround in petsc_matrix_base.h
because nvcc 12.3 with gcc 12.3 complain about accessor being private.
The macro used until now is only defined when compiling device code, not
when nvcc is compiling the host version of the code. This is now fixed
because __CUDACC__ is always defined when nvcc is used.