CMake: do not default-enable taskflow
We simply want the variable to be unset so that we *autodetect* support.

With the bundled variant in place this does exactly what we want -
taskflow is simply enabled by default. If we set the cached variable, on
the other hand, we *force* to have taskflow enabled. This is a problem
in scenarios where bundled libraries cannot be used...